big boi refactor time so that params has more than the model params
