+++ code/main.py
@@ -1,7 +1,9 @@
 import sys
 import pygame
 from Simulation import Simulation
-from DriverModel import Driver, ModelParams
+from DriverModel import Driver
+import Car
+from Road import Road
 
 BLACK = (0, 0, 0)
 WHITE = (255, 255, 255)
@@ -18,12 +20,11 @@ def dist(a, b):
     return np.sqrt((a[0]-b[0])**2 + (a[1]-b[1])**2)
 
 if __name__ == "__main__":
-    a = ModelParams(v_0=30, s_0=2, s_1=0, T=1.6, a=2, b=1.67, delta=4, length=5, thr=0.4, pol=0.5)
-	# v_0 = 125 and 80 km/h, rest are standard values from the paper/traffic-simulation.de
-    car_normal = ModelParams(v_0=34.72, T=1.5, a=0.3, b=3, delta=4, s_0=2, s_1=0, length=5, thr=0.2, pol=0.5)
-    truck_normal = ModelParams(v_0=22.224, T=1.7, a=0.3, b=2, delta=4, s_0=2, s_1=0, length=16.5, thr=0.2, pol=0.5)
+    #v_0 = 125 and 80 km/h, rest are standard values from the paper/traffic-simulation.de
+    car_normal = Car.Params(v_0=34.72, T=1.5, a=0.3, b=3, delta=4, s_0=2, s_1=0, length=5, thr=0.2, pol=0.5, spawn_weight=126090)
+    truck_normal = Car.Params(v_0=22.224, T=1.7, a=0.3, b=2, delta=4, s_0=2, s_1=0, length=16.5, thr=0.2, pol=0.5, spawn_weight=6730)
 
-    sim = Simulation(model_params_list=[(car_normal, 126090), (truck_normal, 6730)], delta_t=DELTA_T, car_frequency=1/1.6)
+    sim = Simulation(params_list=[car_normal, truck_normal], delta_t=DELTA_T, car_frequency=1/1.6)
     data = sim.run(250)
 
     pygame.init()
